Feat: fix product prices

+++ db/migrate/20230517203231_add_price_options_to_product.rb
@@ -1,5 +1,5 @@
 class AddPriceOptionsToProduct < ActiveRecord::Migration[7.0]
   def change
-    add_column :products, :price_options, :integer, array: true, default: []
+    add_column :products, :price_options, :float, array: true, default: []
   end
 end
